+++ system/highlight.rs
@@ -957,7 +957,8 @@ fn classify_terminal(
         return;
     }
 
-    for i in 0..node.child_count() {
+    #[expect(clippy::cast_possible_truncation)]
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             classify_terminal(child, source, lines, categorize);
         }
